test: migrate e2e scenario tests to annotation-DSL feature tier
Delete e2e_code_actions, e2e_rename, and e2e_signature_help — each test
was a loose substring/count assertion that has been replaced by an
expect_test snapshot in the corresponding feature_*.rs file.

Trim e2e_diagnostics, e2e_completion, and e2e_references of every test
whose value was "does this feature exist" rather than "does this
wire-level behaviour hold". Scenario coverage (which symbols, which
diagnostics, which scoping rules) now lives in feature_diagnostics and
feature_references using the caret-annotation DSL, which also asserts
exclusions bidirectionally.

Add #[track_caller] to every check_* helper in TestServer so failures
point at the call site instead of the helper.

Remaining e2e_references tests are those that exercise wire-level
behaviour the annotation DSL cannot express (includeDeclaration flag,
workspace-scan fast path, parallel consistency).
